+++ examples/counter/reducers/index.ts
@@ -2,9 +2,11 @@ import * as Redux from 'redux';
 const { combineReducers } = Redux;
 import { RootState } from '../store';
 import counter from './counter';
+import pathDemo from './path-demo';
 
 const rootReducer = combineReducers<RootState>({
-  counter
+  counter,
+  pathDemo
 });
 
 export default rootReducer;