+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 3.0.6
+
+### Features
+
+Added a 'path' option to `ngRedux.select()` and `@select()`. Now you can
+do stuff like `@select(['foo', 'bar'])` to select `state.foo.bar` into
+an observable.
+
 # 3.0.0
 
 ### Features